implement useTodoReducer custom hook
