+++ src/types.ts
@@ -18,3 +18,8 @@ export interface globalContext {
 export interface Filter {
   filter: string;
 }
+
+// Action types
+export const SET_FILTER = "SET_FILTER";
+export const SET_TODOS = "SET_TODOS";
+export const SET_ISLOADING = "SET_ISLOADING";
